Adds logic to opperator button
Added logic to check if an operator is clicked more than once before
the equals button. If it is, operate on the two values and display it
before letting you do another operation. The logic used for cleaning
the display value in the equals button was extracted into its own
function to make things easier and reduce code duplication.